Generalize trellis-app component
Resolves #99
Resolves #104
