+++ trellis-app-triplestore/src/main/java/org/trellisldp/app/triplestore/RDFConnectionHealthCheck.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.app.health;
+package org.trellisldp.app.triplestore;
 
 import static com.codahale.metrics.health.HealthCheck.Result.healthy;
 import static com.codahale.metrics.health.HealthCheck.Result.unhealthy;
