+++ trellis-app-triplestore/src/test/java/org/trellisldp/app/triplestore/RDFConnectionHealthCheckTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.app.health;
+package org.trellisldp.app.triplestore;
 
 import static org.apache.jena.query.DatasetFactory.wrap;
 import static org.apache.jena.rdfconnection.RDFConnectionFactory.connect;
