+++ trellis-app-triplestore/src/test/java/org/trellisldp/app/triplestore/TrellisAuthorizationTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.app;
+package org.trellisldp.app.triplestore;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static io.dropwizard.testing.ResourceHelpers.resourceFilePath;
@@ -25,16 +25,17 @@ import javax.ws.rs.client.Client;
 
 import org.junit.jupiter.api.AfterAll;
 import org.trellisldp.app.config.TrellisConfiguration;
-import org.trellisldp.test.AbstractApplicationLdpTests;
+import org.trellisldp.test.AbstractApplicationAuthTests;
 
 /**
- * Run LDP-Related Tests.
+ * Authorization tests.
  */
-public class TrellisLdpTest extends AbstractApplicationLdpTests {
+public class TrellisAuthorizationTest extends AbstractApplicationAuthTests {
 
     private static final DropwizardTestSupport<TrellisConfiguration> APP
         = new DropwizardTestSupport<TrellisConfiguration>(TrellisApplication.class,
                 resourceFilePath("trellis-config.yml"),
+                config("auth.basic.usersFile", resourceFilePath("users.auth")),
                 config("binaries", resourceFilePath("data") + "/binaries"),
                 config("mementos", resourceFilePath("data") + "/mementos"),
                 config("namespaces", resourceFilePath("data/namespaces.json")));
@@ -43,7 +44,7 @@ public class TrellisLdpTest extends AbstractApplicationLdpTests {
 
     static {
         APP.before();
-        CLIENT = new JerseyClientBuilder(APP.getEnvironment()).build("test client");
+        CLIENT = new JerseyClientBuilder(APP.getEnvironment()).build("test client2");
         CLIENT.property(CONNECT_TIMEOUT, 5000);
         CLIENT.property(READ_TIMEOUT, 5000);
     }
@@ -56,6 +57,21 @@ public class TrellisLdpTest extends AbstractApplicationLdpTests {
     @Override
     public String getBaseURL() {
         return "http://localhost:" + APP.getLocalPort() + "/";
+    }
+
+    @Override
+    public String getUser1Credentials() {
+        return "acoburn:secret";
+    }
+
+    @Override
+    public String getUser2Credentials() {
+        return "user:password";
+    }
+
+    @Override
+    public String getJwtSecret() {
+        return "secret";
     }
 
     @AfterAll
