+++ trellis-app-triplestore/src/test/resources/config1.yml
@@ -1,0 +1,89 @@
+server:
+  applicationConnectors:
+    - type: http
+      port: 8080
+
+logging:
+  level: WARN
+  loggers:
+    org.trellisldp: INFO
+    io.dropwizard: INFO
+
+binaries: /tmp/trellisData/binaries
+
+mementos: /tmp/trellisData/mementos
+
+baseUrl: http://localhost:8080/
+
+namespaces: /tmp/trellisData/namespaces.json
+
+auth:
+    webac:
+        enabled: true
+        cacheSize: 100
+        cacheExpireSeconds: 10
+    anon:
+        enabled: true
+    basic:
+        enabled: true
+        usersFile: users.auth
+    jwt:
+        enabled: true
+        base64Encoded: false
+        key: secret
+
+defaultName: Trellis
+cacheMaxAge: 86400
+
+assets:
+    icon: "http://example.org/image.icon"
+    js:
+        - "http://example.org/scripts1.js"
+        - "http://example.org/scripts2.js"
+    css:
+        - "http://example.org/styles1.css"
+        - "http://example.org/styles2.css"
+
+notifications:
+    enabled: false
+    type: NONE
+    topicName: "test"
+    connectionString: example.com:1234
+    some.other.value: foo
+
+jsonld:
+    cacheSize: 10
+    cacheExpireHours: 48
+    contextDomainWhitelist: []
+    contextWhitelist:
+        - "http://example.org/context.json"
+
+cors:
+    enabled: true
+    allowOrigin:
+        - "*"
+    allowMethods:
+        - "GET"
+        - "POST"
+        - "PATCH"
+    allowHeaders:
+        - "Content-Type"
+        - "Link"
+    exposeHeaders:
+        - "Link"
+        - "Location"
+    maxAge: 180
+    allowCredentials: true
+
+binaryHierarchyLevels: 2
+binaryHierarchyLength: 1
+
+cassandraAddress: my.cluster.node
+cassandraPort: 245993
+
+extraConfigValues:
+    one: true
+    two:
+      - "value1"
+      - "value2"
+
