+++ trellis-app-triplestore/src/test/resources/data/namespaces.json
