+++ trellis-app-triplestore/src/test/resources/users.auth
@@ -1,0 +1,4 @@
+# A comment
+other : pass
+acoburn : secret : https://people.apache.org/~acoburn/#i
+user : password : https://madison.example.com/profile/#me
