+++ trellis-app/build.gradle
@@ -1,58 +1,38 @@
-apply plugin: 'java'
-apply plugin: 'distribution'
-apply plugin: 'application'
-
 description = 'Trellis Application'
-mainClassName = 'org.trellisldp.app.TrellisApplication'
+
+ext {
+    moduleName = 'org.trellisldp.app'
+}
 
 dependencies {
-    compile group: 'commons-codec', name: 'commons-codec', version: commonsCodecVersion
-    compile group: 'javax.activation', name: 'javax.activation-api', version: activationApiVersion
-    compile group: 'javax.jms', name: 'javax.jms-api', version: jmsApiVersion
-    compile group: 'io.dropwizard', name: 'dropwizard-auth', version: dropwizardVersion
-    compile group: 'io.dropwizard', name: 'dropwizard-core', version: dropwizardVersion
-    compile group: 'io.dropwizard', name: 'dropwizard-metrics', version: dropwizardVersion
-    compile group: 'io.dropwizard', name: 'dropwizard-http2', version: dropwizardVersion
-    compile group: 'io.jsonwebtoken', name: 'jjwt', version: jwtVersion
-    compile group: 'org.apache.activemq', name: 'activemq-client', version: activeMqVersion
-    compile group: 'org.apache.commons', name: 'commons-rdf-jena', version: commonsRdfVersion
-    compile group: 'org.apache.jena', name: 'jena-osgi', version: jenaVersion
-    compile group: 'org.apache.jena', name: 'jena-tdb2', version: jenaVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
+    api group: 'io.dropwizard', name: 'dropwizard-auth', version: dropwizardVersion
+    api group: 'io.dropwizard', name: 'dropwizard-core', version: dropwizardVersion
+    api project(':trellis-api')
 
-    compile project(':trellis-constraint-rules')
-    compile project(':trellis-io-jena')
-    compile project(':trellis-id')
-    compile project(':trellis-api')
-    compile project(':trellis-http')
-    compile project(':trellis-vocabulary')
-    compile project(':trellis-file')
-    compile project(':trellis-namespaces')
-    compile project(':trellis-agent')
-    compile project(':trellis-audit')
-    compile project(':trellis-event-serialization')
-    compile project(':trellis-webac')
-    compile project(':trellis-triplestore')
-    compile project(':trellis-jms')
-    compile project(':trellis-kafka')
+    implementation group: 'io.jsonwebtoken', name: 'jjwt', version: jwtVersion
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
+    implementation project(':trellis-http')
+    implementation project(':trellis-vocabulary')
+    implementation project(':trellis-agent')
+    implementation project(':trellis-webac')
 
-    runtime group: 'javax.xml.bind', name: 'jaxb-api', version: jaxbVersion
+    testRuntime group: 'javax.xml.bind', name: 'jaxb-api', version: jaxbVersion
+    testRuntime group: 'javax.activation', name: 'javax.activation-api', version: activationApiVersion
+    testRuntime group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: junitVersion
 
-    testCompile group: 'io.dropwizard', name: 'dropwizard-client', version: dropwizardVersion
-    testCompile group: 'io.dropwizard', name: 'dropwizard-testing', version: dropwizardVersion
-    testCompile group: 'org.apache.activemq', name: 'activemq-broker', version: activeMqVersion
-    testCompile group: 'org.apiguardian', name: 'apiguardian-api', version: apiguardianVersion
-    testCompile group: 'org.apache.tamaya', name: 'tamaya-core', version: tamayaVersion
-    testCompile group: 'org.awaitility', name: 'awaitility', version: awaitilityVersion
-    testCompile group: 'ch.qos.logback', name: 'logback-classic', version: logbackVersion
-    testCompile group: 'org.mockito', name: 'mockito-core', version: mockitoVersion
-    testCompile project(':trellis-test')
+    testImplementation group: 'io.dropwizard', name: 'dropwizard-client', version: dropwizardVersion
+    testImplementation group: 'io.dropwizard', name: 'dropwizard-testing', version: dropwizardVersion
+    testImplementation group: 'org.apache.commons', name: 'commons-rdf-jena', version: commonsRdfVersion
+    testImplementation group: 'org.apache.jena', name: 'jena-osgi', version: jenaVersion
+    testImplementation group: 'org.apache.jena', name: 'jena-tdb2', version: jenaVersion
+    testImplementation group: 'org.apiguardian', name: 'apiguardian-api', version: apiguardianVersion
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: junitVersion
+    testImplementation group: 'org.awaitility', name: 'awaitility', version: awaitilityVersion
+    testImplementation group: 'ch.qos.logback', name: 'logback-classic', version: logbackVersion
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: mockitoVersion
+    testImplementation project(':trellis-triplestore')
+    testImplementation project(':trellis-io-jena')
+    testImplementation project(':trellis-file')
+    testImplementation project(':trellis-id')
 }
-
-task copyDistTask(type: Copy) {
-    from 'LICENSE'
-    into 'src/dist'
-}
-
-assembleDist.dependsOn copyDistTask
